backend/requirements.txt
fastapi
uvicorn
pymupdf
pillow
python-multipart
python-docx
docx2pdf